Enforce Unix newlines through rustfmt
This was previously done before 5984e0d: "Add integration tests for macOS
and Windows (#2)", which (so I thought) forced a change due to
Travis's repository cloning on Windows changing the newline style,
causing rustfmt to fail on Windows consistently. Newline conversions on
clone can be disabled according to
https://docs.travis-ci.com/user/customizing-the-build/#git-end-of-line-conversion-control,
allowing for enforcement of Unix newlines once again.