Fix 4a9308d: "Use stable clippy & rustfmt when nightly fails (#5)" (#8)
4a9308d attempted to solve the problem of nightly sometimes not having
clippy or rustfmt available by installing the stable versions of these
tools as a backup. What it neglected to address was that the `cargo
clippy` and `cargo fmt` commands used *didn't try to use the stable
versions on nightly toolchains*, rendering 4a9308d effectively
uneffective. Fancy that.

This commit *always* uses the stable versions of clippy and rustfmt on
*all* toolchains to prevent errors. We don't need to test nightly rust
tools; we need to test this repository.

The `before_install` section was changed to `before_script` as well to
reflect when the clippy and rustfmt installs are actually needed.