Use nightly archive version of rustfmt (#15)
There are a few conflicts between my preferred coding style and the
options available in stable rustfmt, so rather than live with it,
rustfmt will now run on a specific nightly archive. It does not run on
the latest nightly because it is not necessarily available in any given
nightly.