Add debug lines for demonstration purposes
The intent is that this commit will be reverted when proper world
generation is implemented.

The integer_division clippy lint was removed in the process of
implementing this feature. It is simply too strict for most use cases,
and with the help of macros like `const_assert_eq!`, there is little
reason to keep it around. This removal should stay when the rest of this
commit is reverted.