Don't use squash merge strategy for bors
As ugly as merge commits may look to me, they have the following major
benefits:

 - Preserving commit signatures (if applicable)

 - Preserving detailed commit messages

It should be noted that both of these are preserved in GitHub's
"history" through past PR's. However, these things should really be held
in the commit history for portability reasons (e.g. mirrors) and the
commit history is meant to hold them anyway.