Update Cargo.lock
This is to "bootstrap" dependabot in a way. Since the lockfile hasn't
been updated in quite a while, dependabot would have to make 30-ish PR's
to make all dependencies up-to-date. This "resets" us to the most
up-to-date dependencies to prevent dependabot from clogging up CI. All
further updates to the lockfile should be done by dependabot.