Allow lockfile updates (again)
I misunderstood the dependabot documentation where it said

> By default all dependencies that are explicitly defined in a manifest
> or lock file are kept up to date.

to mean that Cargo.lock would be updated by default. This appears not to
be the case, and the "dependency-type" option appears to default to
"direct" rather than "all". Updating Cargo.lock only worked previously
because dependabot-preview was being used at the time. The error occured
in translation from dependabot-preview's online configuration to the
YAML configuration.