Bump amethyst from 0.15.0 to 0.15.1 in Cargo.toml
In the middle of the recent migration from dependabot-preview to
dependabot proper, Amethyst released version 0.15.1. Cargo.lock got
updated automatically, but in the slight hiccup in the migration
Cargo.toml wasn't updated and won't be updated by dependabot (at least
until the next version of amethyst is released. I don't know exactly how
dependabot works at this point in time, but I think dependabot won't
update Cargo.toml because Cargo.lock was updated before it, making
dependabot think amethyst is up-to-date). Whatever the case, we should
do this explicitly, and we'll hope dependabot does what we want when the
next version of amethyst is released.